Normalise link anchors to match generated heading ids
